src/requirements/base.txt
-e git://github.com/marteinn/atomicpress@develop#egg=atomicpress
python-dotenv==0.5.0
